requirements.txt
Flask==3.0.1
Flask-Cors==4.0.0
stripe==10.11.0
python-dotenv==1.0.0
sendgrid==6.10.0
psycopg2
